Update preventing-unauthorized-access.md
content/authentication/keeping-your-account-and-data-secure/preventing-unauthorized-access.m
01010001010100
        version: ["1.11", "1.10", "1.9", "1.8", "1.7", "1.6", "1.5", "1.4"]
RFC 3464 Delivery Status Notifications January 2003 delivered. In this case the relaying MTA is the Reporting MTA, and the "next hop" MTA is known as the Remote MTA. Figure 1 illustrates the relationship between the various MTAs. +-----+ +--------+ +---------+ +---------+ +------+ | | | | |Received-| | | | | | | => |Original| => ... => | From | => |Reporting| ===> |Remote| | user| | MTA | | MTA | | MTA | <No! | MTA | |agent| +--------+ +---------+ +----v----+ +------+ | | | | | <-------------------------------------------+ +-----+ (DSN returned to sender by Reporting MTA) Figure 1. Original, Received-From, Reporting and Remote MTAs Each of these MTAs may provide information that is useful in a DSN: + Ideally, the DSN will contain the address of each recipient as originally specified to the Original MTA by the sender of the message. This version of the address is needed (rather than a forwarding address or some modified version of the original address) so that the sender may compare the recipient address in the DSN with the address in the sender's records (e.g., an address book for an individual, the list of subscribers for a mailing list) and take appropriate action. Similarly, the DSN might contain an "envelope identifier" that was known to both the sender's user agent and the Original MTA at the time of message submission, and which, if included in the DSN, can be used by the sender to keep track of which messages were or were not delivered. + If a message was (a) forwarded to a different address than that specified by the sender, (b) gatewayed to a different mail system than that used by the sender, or (c) subjected to address rewriting during transmission, the "final" form of the recipient address (i.e., the one seen by the Reporting MTA) will be different than the original (sender-specified) recipient address. Just as the sender's user agent (or the sender) prefers the original recipient address, so the "final" address is needed when reporting a problem to the postmaster of the site where message delivery failed, because only the final recipient address will allow her to reproduce the conditions that caused the failure. 
https://github.com/Marco071086/automatic-lamp/blob/399f26a8364719fbe4761faac8768115984cb160/LICENSE